add walk rate; fix panic

+++ src/model.rs
@@ -1,18 +1,18 @@
 
 pub struct Model {
-    pub x: u32,
-    pub y: u32,
-    pub vx: u32,
-    pub vy: u32,
+    pub x: f32,
+    pub y: f32,
+    pub vx: f32,
+    pub vy: f32,
 }
 
 impl Model {
     pub fn new() -> Self {
         Model {
-            x: 0,
-            y: 0,
-            vx: 0,
-            vy: 0,
+            x: 0.0,
+            y: 0.0,
+            vx: 0.0,
+            vy: 0.0,
         }
     }
 }
